🗺️ Atlas: [Extract Telemetry Modules]
- Extracted the 6 distinct telemetry channels from the monolithic `lib.rs` into specialized submodules (`bytecode_cost.rs`, `object_lineage.rs`, etc.).
- `lib.rs` now acts as a clean facade, maintaining backwards compatibility.
- Moved and isolated serde helper functions to a private `helpers.rs` module.
