+++ .jules/atlas.md
@@ -16,3 +16,6 @@
 **Extracting Host I/O from duke-gc Blob**
 **Tangle:** The `duke-gc` crate's `lib.rs` file had become a massive Blob anti-pattern (over 2500 lines). It combined complex generational garbage collection logic (Eden space, copying collector, mark-sweep) with low-level host OS abstractions like file descriptors, sockets, and child process management.
 **Blueprint:** Extracted the host I/O management structures (`HostFileHandle`, `HostProcessHandle`, `SpawnedProcessIds`) and all related native host methods on `Heap` into a separate `host.rs` module. The JVM host interactions are now neatly encapsulated without polluting the garbage collection logic.
+**[Refactoring the Duke Telemetry Blob]
+**Tangle:** The `crates/duke-telemetry/src/lib.rs` file was a massive Blob anti-pattern (over 1000 lines). It combined 6 distinct telemetry channels (bytecode cost, object lineage, class initialization, exception flow, dispatch resolution, and native boundary) along with the central store and formatting tools.
+**Blueprint:** Extracted the 6 individual channels and the serde formatting helpers into distinct submodules (`bytecode_cost.rs`, `object_lineage.rs`, etc.). `lib.rs` was refactored into a clean facade that re-exports the individual channels and acts as the singular `TelemetryStore` entrypoint, maintaining backwards compatibility while restoring domain responsibility separation.
