+++ crates/duke-telemetry/src/helpers.rs
@@ -1,0 +1,52 @@
+#[cfg(feature = "telemetry")]
+pub mod ser_helpers {
+    use std::collections::HashMap;
+
+    use serde::Serialize;
+
+    /// Core helper: serialize any `HashMap<K, V>` by formatting each key with `key_fn`.
+    fn keyed_map<K, V, S, F>(map: &HashMap<K, V>, ser: S, key_fn: F) -> Result<S::Ok, S::Error>
+    where
+        K: Eq + std::hash::Hash,
+        V: Serialize,
+        S: serde::Serializer,
+        F: Fn(&K) -> String,
+    {
+        let string_map: HashMap<String, &V> = map.iter().map(|(k, v)| (key_fn(k), v)).collect();
+        string_map.serialize(ser)
+    }
+
+    /// `HashMap<(class, method, pc), V>` → `"class::method@pc"`.
+    pub fn site3<V: Serialize, S: serde::Serializer>(
+        map: &HashMap<(String, String, usize), V>,
+        ser: S,
+    ) -> Result<S::Ok, S::Error> {
+        keyed_map(map, ser, |(c, m, pc)| format!("{c}::{m}@{pc}"))
+    }
+
+    /// `HashMap<(class, cp_idx), V>` → `"class@cp"`.
+    pub fn site2_u16<V: Serialize, S: serde::Serializer>(
+        map: &HashMap<(String, u16), V>,
+        ser: S,
+    ) -> Result<S::Ok, S::Error> {
+        keyed_map(map, ser, |(c, cp)| format!("{c}@{cp}"))
+    }
+
+    /// `HashMap<(class, method), V>` → `"class::method"`.
+    pub fn pair_str<V: Serialize, S: serde::Serializer>(
+        map: &HashMap<(String, String), V>,
+        ser: S,
+    ) -> Result<S::Ok, S::Error> {
+        keyed_map(map, ser, |(c, m)| format!("{c}::{m}"))
+    }
+
+    /// Serialize `HashSet<String>` as a sorted `Vec<String>` for deterministic output.
+    pub fn sorted_set<S: serde::Serializer>(
+        set: &std::collections::HashSet<String>,
+        ser: S,
+    ) -> Result<S::Ok, S::Error> {
+        let mut v: Vec<&String> = set.iter().collect();
+        v.sort();
+        v.serialize(ser)
+    }
+}
